+++ src/org/labkey/test/components/experiment/UserActivateDeactivateConfirmationDialog.java
@@ -1,8 +1,7 @@
-package org.labkey.test.components.ui.user;
+package org.labkey.test.components.experiment;
 
 import org.labkey.test.WebDriverWrapper;
 import org.labkey.test.components.bootstrap.ModalDialog;
-import org.labkey.test.components.ui.ConfirmationDialog;
 import org.labkey.test.pages.LabKeyPage;
 
 import java.util.function.Supplier;
@@ -10,16 +9,16 @@ import java.util.function.Supplier;
 /**
  * Wraps the UI component defined by 'internal/components/user/UserActivateChangeConfirmModal.tsx'
  */
-public class UserActivateChangeConfirmModal<SourcePage extends WebDriverWrapper, ConfirmPage extends LabKeyPage<?>> extends ConfirmationDialog<SourcePage, ConfirmPage>
+public class UserActivateDeactivateConfirmationDialog<SourcePage extends WebDriverWrapper, ConfirmPage extends LabKeyPage<?>> extends ConfirmationDialog<SourcePage, ConfirmPage>
 {
     private final boolean _activate;
 
-    public UserActivateChangeConfirmModal(boolean activate, SourcePage sourcePage)
+    public UserActivateDeactivateConfirmationDialog(boolean activate, SourcePage sourcePage)
     {
         this(activate, sourcePage, () -> null);
     }
 
-    protected UserActivateChangeConfirmModal(boolean activate, SourcePage sourcePage, Supplier<ConfirmPage> confirmPageSupplier)
+    protected UserActivateDeactivateConfirmationDialog(boolean activate, SourcePage sourcePage, Supplier<ConfirmPage> confirmPageSupplier)
     {
         super(new ModalDialog.ModalDialogFinder(sourcePage.getDriver()).withTitle("user"), sourcePage, confirmPageSupplier);
         _activate = activate;
